RTP uplifting frame reordering fixes

+++ src/org/jitsi/service/neomedia/MediaStream.java
@@ -520,10 +520,10 @@ public interface MediaStream
      * @param ssrcTargetRTX the SSRC into which the rtxGroups will be rewritten.
      */
     public void configureSSRCRewriting(
-        final Set<Integer> ssrcGroup, final Integer ssrcTargetPrimary,
-        final Map<Integer, Byte> ssrc2fec,
-        final Map<Integer, Byte> ssrc2red,
-        final Map<Integer, Integer> rtxGroups, final Integer ssrcTargetRTX);
+        final Set<Long> ssrcGroup, final Long ssrcTargetPrimary,
+        final Map<Long, Byte> ssrc2fec,
+        final Map<Long, Byte> ssrc2red,
+        final Map<Long, Long> rtxGroups, final Long ssrcTargetRTX);
 
     /**
      * Gets the {@link RawPacketCache} which (optionally) caches outgoing
